+++ extensions/CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+## 5.1.0 - 2023-05-04
+
 ### Added
 
 - `checkPackages` function to check a list of packages against the default policy